Move IllegalValueException
